Fixes according to review
